LayoutTests/js/dom/domjit-accessor-monomorphic-expected.txt

Test DOMJIT accessors work in monomorphic call sites.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS func("firstChild", document.getElementById("target"), document.getElementById("firstChild")) is true
PASS func("lastChild", document.getElementById("target"), document.getElementById("lastChild")) is true
PASS func("nextSibling", document.getElementById("target"), document.getElementById("nextSibling")) is true
PASS func("previousSibling", document.getElementById("target"), document.getElementById("previousSibling")) is true
PASS func("parentNode", document.getElementById("target"), document.getElementById("parentNode")) is true
PASS successfullyParsed is true

TEST COMPLETE
